src/dataflow/dfl_delivery_sensor/requirements.txt
apache_beam[gcp]==2.66.0
google-cloud-bigquery==3.36.0
